+++ marginaleffects/uncertainty.py
@@ -5,7 +5,7 @@ import polars as pl
 import scipy.stats as stats
 
 
-def get_jacobian(func, coefs):
+def get_jacobian(func, coefs, eps_vcov=None):
     # forward finite difference (faster)
     if coefs.ndim == 2:
         if isinstance(coefs, np.ndarray):
@@ -15,7 +15,10 @@ def get_jacobian(func, coefs):
         baseline = func(coefs)["estimate"].to_numpy()
         jac = np.empty((baseline.shape[0], len(coefs_flat)), dtype=np.float64)
         for i, xi in enumerate(coefs_flat):
-            h = max(abs(xi) * np.sqrt(np.finfo(float).eps), 1e-10)
+            if eps_vcov is not None:
+                h = eps_vcov
+            else:
+                h = max(abs(xi) * np.sqrt(np.finfo(float).eps), 1e-10)
             dx = np.copy(coefs_flat)
             dx[i] = dx[i] + h
             tmp = dx.reshape(coefs.shape)
@@ -25,7 +28,10 @@ def get_jacobian(func, coefs):
         baseline = func(coefs)["estimate"].to_numpy()
         jac = np.empty((baseline.shape[0], len(coefs)), dtype=np.float64)
         for i, xi in enumerate(coefs):
-            h = max(abs(xi) * np.sqrt(np.finfo(float).eps), 1e-10)
+            if eps_vcov is not None:
+                h = eps_vcov
+            else:
+                h = max(abs(xi) * np.sqrt(np.finfo(float).eps), 1e-10)
             dx = np.copy(coefs)
             dx[i] = dx[i] + h
             jac[:, i] = (func(dx)["estimate"].to_numpy() - baseline) / h